Reorders aria roles in tab actions use case
