+++ src/tabs/tab-header-bar.tsx
@@ -28,8 +28,8 @@ import { getAllFocusables } from '../internal/components/focus-lock/utils';
 import { nodeBelongs } from '../internal/utils/node-belongs';
 import { ButtonProps } from '../button/interfaces';
 
-const tabSelector = '[role="tab"]';
-const activeTabSelector = '[role="tab"][aria-selected="true"]';
+const tabSelector = `.${styles['tabs-tab-link']}`;
+const activeTabSelector = `${tabSelector}.${styles['tabs-tab-active']}`;
 
 function dismissButton(dismissLabel: TabsProps.Tab['dismissLabel'], onDismiss: TabsProps.Tab['onDismiss']) {
   return (
@@ -72,6 +72,19 @@ export function TabHeaderBar({
   const [horizontalOverflow, setHorizontalOverflow] = useState(false);
   const [inlineStartOverflow, setInlineStartOverflow] = useState(false);
   const [inlineEndOverflow, setInlineEndOverflow] = useState(false);
+  const showTabActionAttributes = tabs.some(tab => tab.action || tab.dismissible);
+  const tabsWithActionsAriaRoleDescription = 'Tabs with Actions';
+  const tabActionAttributes = showTabActionAttributes
+    ? {
+        role: 'application',
+        'aria-roledescription': i18n(
+          'i18nStrings.tabsWithActionsAriaRoleDescription',
+          tabsWithActionsAriaRoleDescription
+        ),
+      }
+    : {
+        role: 'tablist',
+      };
 
   useEffect(() => {
     if (headerBarRef.current) {
@@ -244,7 +257,6 @@ export function TabHeaderBar({
         onUnregisterFocusable={onUnregisterFocusable}
       >
         <ul
-          role="tablist"
           className={styles['tabs-header-list']}
           aria-label={ariaLabel}
           aria-labelledby={ariaLabelledby}
@@ -253,6 +265,7 @@ export function TabHeaderBar({
           onKeyDown={onKeyDown}
           onFocus={onFocus}
           onBlur={onBlur}
+          {...tabActionAttributes}
         >
           {tabs.map(renderTabHeader)}
         </ul>
@@ -275,6 +288,8 @@ export function TabHeaderBar({
 
   function renderTabHeader(tab: TabsProps.Tab) {
     const { dismissible, dismissLabel, action, onDismiss } = tab;
+    const hasActionOrDismissible = action || dismissible;
+    const groupRole = hasActionOrDismissible ? 'group' : undefined;
 
     const clickTab = (event: React.MouseEvent) => {
       if (tab.disabled) {
@@ -322,13 +337,16 @@ export function TabHeaderBar({
 
     const commonProps: (JSX.IntrinsicElements['a'] | JSX.IntrinsicElements['button']) & { 'data-testid': string } = {
       className: classes,
-      role: 'tab',
-      'aria-selected': activeTabId === tab.id,
+      role: hasActionOrDismissible ? 'button' : 'tab',
       'aria-controls': `${idNamespace}-${tab.id}-panel`,
       'data-testid': tab.id,
       id: getTabElementId({ namespace: idNamespace, tabId: tab.id }),
       children: <span className={styles['tabs-tab-label']}>{tab.label}</span>,
     };
+
+    if (!hasActionOrDismissible) {
+      commonProps['aria-selected'] = activeTabId === tab.id;
+    }
 
     if (tab.disabled) {
       commonProps['aria-disabled'] = 'true';
@@ -368,7 +386,7 @@ export function TabHeaderBar({
         role="presentation"
         key={tab.id}
       >
-        <div className={tabHeaderContainerClasses} role="group" aria-labelledby={commonProps.id}>
+        <div className={tabHeaderContainerClasses} role={groupRole}>
           <TabTrigger ref={setElement} tab={tab} elementProps={commonProps} />
           {action && <span className={styles['tabs-tab-action']}>{action}</span>}
           {dismissible && (
@@ -385,7 +403,10 @@ const TabTrigger = forwardRef(
     {
       tab,
       elementProps,
-    }: { tab: TabsProps.Tab; elementProps: React.HTMLAttributes<HTMLAnchorElement | HTMLButtonElement> },
+    }: {
+      tab: TabsProps.Tab;
+      elementProps: React.HTMLAttributes<HTMLAnchorElement | HTMLButtonElement>;
+    },
     ref: React.Ref<HTMLElement>
   ) => {
     const refObject = useRef<HTMLElement>(null);
